Tuesday in class

+++ src/hoverboard.java
@@ -1,9 +1,9 @@
 public class hoverboard extends skateboard {
-        int maxHover;
+        double maxHover;
         int maxPower;
 
 
-    public hoverboard(String inBrand, double inSpeed, int inPassengers, double inCargo, int weight, double deckSize, String material, boolean wheels, int maxHover, int maxPower) {
+    public hoverboard(String inBrand, double inSpeed, int inPassengers, double inCargo, int weight, double deckSize, String material, boolean wheels, double maxHover, int maxPower) {
         super(inBrand, inSpeed, inPassengers, inCargo, weight, deckSize, material);
             this.maxHover = maxHover;
             this.maxPower = maxPower;
